+++ eth_tx_trie_test.go
@@ -7,14 +7,10 @@ import (
 
 func TestTxTriesInBlockBodyRlpParsing(t *testing.T) {
 	fi, err := os.Open("test_data/eth-block-body-rlp-999999")
-	if err != nil {
-		t.Fatal(err)
-	}
+	checkError(err, t)
 
 	_, _, output, err := FromBlockRLP(fi)
-	if err != nil {
-		t.Fatal(err)
-	}
+	checkError(err, t)
 
 	if len(output) != 13 {
 		t.Fatal("Wrong number of obtained tx trie nodes")