+++ plugin/CHANGELOG.md
@@ -7,8 +7,14 @@
     * `eth-block`
       * Accept input in RLP for both block headers or block bodies. (as `raw`).
       * Accept input in JSON.
+      * Display common hashes as well as `cid`s on JSON Marshaling.
   * Work on the output:
     * When queried, `eth-block` always responds a block header in RLP.
+    * Add resolving (or add tests) for the following elements:
+      * `bloom`
+      * `coinbase`
+      * `parent`
+  * Test coverage is `52.7%`.
 
 ## `0.0.2`